refactor: change function name to `rangeToSQLInterval` and rename files
refactor: change function name to `rangeToSQLInterval`

refactor: rename files

chore: lint
